+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/autocrafting/craftingmonitor/CraftingMonitorElementFluidRender.java
@@ -25,12 +25,12 @@ public class CraftingMonitorElementFluidRender implements ICraftingMonitorElemen
     public static final String ID = "fluid_render";
 
     private final FluidStack stack;
-    private int stored;
-    private int processing;
-    private int scheduled;
-    private int crafting;
+    private long stored;
+    private long processing;
+    private long scheduled;
+    private long crafting;
 
-    public CraftingMonitorElementFluidRender(FluidStack stack, int stored, int processing, int scheduled, int crafting) {
+    public CraftingMonitorElementFluidRender(FluidStack stack, long stored, long processing, long scheduled, long crafting) {
         this.stack = stack;
         this.stored = stored;
         this.processing = processing;
@@ -101,32 +101,32 @@ public class CraftingMonitorElementFluidRender implements ICraftingMonitorElemen
     }
 
     @Override
-    public int getStored() {
+    public long getStored() {
         return stored;
     }
 
     @Override
-    public int getProcessing() {
+    public long getProcessing() {
         return processing;
     }
 
     @Override
-    public int getScheduled() {
+    public long getScheduled() {
         return scheduled;
     }
 
     @Override
-    public int getCrafting() {
+    public long getCrafting() {
         return crafting;
     }
 
     @Override
     public void write(ByteBuf buf) {
         ByteBufUtils.writeTag(buf, stack.writeToNBT(new NBTTagCompound()));
-        buf.writeInt(stored);
-        buf.writeInt(processing);
-        buf.writeInt(scheduled);
-        buf.writeInt(crafting);
+        buf.writeLong(stored);
+        buf.writeLong(processing);
+        buf.writeLong(scheduled);
+        buf.writeLong(crafting);
     }
 
     @Override
