+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/autocrafting/registry/CraftingTaskFactory.java
@@ -15,7 +15,7 @@ public class CraftingTaskFactory implements ICraftingTaskFactory {
 
     @Nonnull
     @Override
-    public MasterCraftingTask create(INetwork network, ICraftingRequestInfo requested, int quantity, ICraftingPattern pattern) {
+    public MasterCraftingTask create(INetwork network, ICraftingRequestInfo requested, long quantity, ICraftingPattern pattern) {
         return new MasterCraftingTask(network, requested, quantity, pattern);
     }
