+++ src/main/java/com/raoulvdberge/refinedstorage/render/constants/ConstantsDisk.java
@@ -9,10 +9,10 @@ public final class ConstantsDisk {
 
     public static final int DISK_NEAR_CAPACITY_THRESHOLD = 75;
 
-    public static int getDiskState(int stored, int capacity) {
+    public static int getDiskState(long stored, long capacity) {
         if (stored == capacity) {
             return DISK_STATE_FULL;
-        } else if ((int) ((float) stored / (float) capacity * 100F) >= DISK_NEAR_CAPACITY_THRESHOLD) {
+        } else if ((long) ((double) stored / (double) capacity * 100F) >= DISK_NEAR_CAPACITY_THRESHOLD) {
             return DISK_STATE_NEAR_CAPACITY;
         } else {
             return DISK_STATE_NORMAL;
